Copilot pull request changes + refactor
